refactored list
